+++ src/arclet/alconna/action.py
@@ -8,16 +8,28 @@ class ActType(IntEnum):
     STORE = 0
     """无 Args 时, 仅存储一个值, 默认为 Ellipsis; 有 Args 时, 后续的解析结果会覆盖之前的值"""
     APPEND = 1
-    """无 Args 时, 将多个值存为列表, 默认为 Ellipsis; 有 Args 时, 每个解析结果会追加到列表中"""
+    """无 Args 时, 将多个值存为列表, 默认为 Ellipsis; 有 Args 时, 每个解析结果会追加到列表中
+    
+    当存在默认值并且不为列表时, 会自动将默认值变成列表, 以保证追加的正确性
+    """
     COUNT = 2
-    """无 Args 时, 计数器加一; 有 Args 时, 表现与 STORE 相同"""
+    """无 Args 时, 计数器加一; 有 Args 时, 表现与 STORE 相同
+    
+    当存在默认值并且不为数字时, 会自动将默认值变成 1, 以保证计数器的正确性
+    """
 
 
-@dataclass
+@dataclass(eq=True)
 class Action:
     """节点触发的动作"""
     type: ActType
     value: Any
+
+    def __post_init__(self):
+        if self.type == ActType.APPEND and not isinstance(self.value, list):
+            self.value = [self.value]
+        elif self.type == ActType.COUNT and not isinstance(self.value, int):
+            self.value = 1
 
 
 store = Action(ActType.STORE, Ellipsis)