+++ src/arclet/alconna/builtin.py
@@ -18,9 +18,7 @@ MISSING = _MISSING_TYPE()
 class _SetDefault(ArparmaBehavior):
     _default: Any = field(default=MISSING)
     _default_factory: Callable | _MISSING_TYPE = field(default=MISSING)
-    arg: str | None = field(default=None)
-    option: str | None = field(default=None)
-    subcommand: str | None = field(default=None)
+    path: str | None = field(default=None)
 
     @property
     def default(self):
@@ -28,58 +26,28 @@ class _SetDefault(ArparmaBehavior):
             return self._default
         if callable(self._default_factory):
             return self._default_factory()
-        raise BehaveCancelled('cannot specify both value and factory')
+        raise BehaveCancelled("cannot specify both value and factory")
 
     def operate(self, interface: Arparma):
-        if not self.option and not self.subcommand:
+        if not self.path:
             raise BehaveCancelled
-        if self.arg and self.arg not in interface.other_args:
-            self.update(interface, f"other_args.{self.arg}", self.default)
-        if self.option and self.subcommand is None:
-            if not interface.query(f"options.{self.option}"):
-                self.update(
-                    interface, f"options.{self.option}",
-                    OptionResult(None, {self.arg: self.default}) if self.arg else OptionResult(self.default)
-                )
-            elif self.arg and not interface.query(f"options.{self.option}.{self.arg}"):
-                self.update(interface, f"options.{self.option}.{self.arg}", self.default)
-        if self.subcommand and self.option is None:
-            if not interface.query(f"subcommands.{self.subcommand}"):
-                self.update(
-                    interface, f"subcommands.{self.subcommand}",
-                    SubcommandResult(None, {self.arg: self.default}) if self.arg else SubcommandResult(self.default)
-                )
-            elif self.arg and not interface.query(f"subcommands.{self.subcommand}.{self.arg}"):
-                self.update(interface, f"subcommands.{self.subcommand}.{self.arg}", self.default)
-        if self.option and self.subcommand:
-            if not interface.query(f"subcommands.{self.subcommand}.options.{self.option}"):
-                self.update(
-                    interface, f"subcommands.{self.subcommand}.options.{self.option}",
-                    OptionResult(None, {self.arg: self.default}) if self.arg else OptionResult(self.default)
-                )
-            elif self.arg and not interface.query(f"subcommands.{self.subcommand}.options.{self.option}.{self.arg}"):
-                self.update(interface, f"subcommands.{self.subcommand}.options.{self.option}.{self.arg}", self.default)
+        def_val = self.default
+        if not interface.query(self.path):
+            self.update(interface, self.path, def_val)
 
 
 @overload
-def set_default(
-    *, value: Any, arg: str | None = None, option: str | None = None, subcommand: str | None = None,
-) -> _SetDefault:
+def set_default(*, value: Any, path: str,) -> _SetDefault:
     ...
 
 
 @overload
-def set_default(
-    *, factory: Callable[..., Any], arg: str | None = None, option: str | None = None, subcommand: str | None = None,
-) -> _SetDefault:
+def set_default(*, factory: Callable[..., Any], path: str,) -> _SetDefault:
     ...
 
 
 def set_default(
-    *,
-    value: Any = MISSING,
-    factory: Callable[..., Any] = MISSING,
-    arg: str | None = None, option: str | None = None, subcommand: str | None = None,
+    *, value: Any = MISSING, factory: Callable[..., Any] = MISSING, path: str | None = None,
 ) -> _SetDefault:
     """
     设置一个选项的默认值, 在无该选项时会被设置
@@ -89,11 +57,9 @@ def set_default(
     Args:
         value (Any): 默认值
         factory (Callable[..., Any]): 默认值工厂
-        arg (str | None): 参数名
-        option (str | None): 选项名
-        subcommand (str | None): 子命令名
+        path: str: 参数路径
     """
     if value is not MISSING and factory is not MISSING:
-        raise ValueError('cannot specify both value and factory')
+        raise ValueError("cannot specify both value and factory")
 
-    return _SetDefault(value, factory, arg, option, subcommand)
+    return _SetDefault(value, factory, path)